# service_auth/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Authentication service
  description: Authentication service
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Role:
      type: object
      description: Role model
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    RolesList:
      type: array
      items:
        $ref: '#/components/schemas/Role'
    RoleNew:
      type: object
      description: Model for creating new role
      required:
        - name
      properties:
        name:
          type: string
...
    ApiErrorResponse:
      type: object
      description: Error response model
      required:
        - msg
      properties:
        msg:
          type: string
    TokensResponse:
      type: object
      description: Access and refresh tokens model
      required:
        - access_token
        - refresh_token
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
    UserAction:
      type: object
      description: User actions model (history of logins)
      required:
        - id
        - user_id
        - device_name
        - timestamp
        - action
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        device_name:
          type: string
        timestamp:
          type: string
          format: date-time
        action:
          type: string
    UserActionsList:
      type: array
      items:
        $ref: '#/components/schemas/UserAction'
    UserSession:
      type: object
      description: User session model (active sessions)
      required:
        - device_name
        - remote_ip
        - login_at
        - active_at
      properties:
        device_name:
          type: string
        remote_ip:
          type: string
        login_at:
          type: string
          format: date-time
        active_at:
          type: string
          format: date-time
    UserSessionsList:
      type: array
      items:
        $ref: '#/components/schemas/UserSession'
  responses:
    200-JWTTokens:
      description: Response with access and refresh token
      headers:
        Set-Cookie:
          description: refresh token in cookie
          schema:
            type: string
        #Set-Cookie:
        #  description: csrf token in cookie
        #  schema:
        #    type: string
        WWW-Authenticate:
          description: access token
          schema:
            type: string
      # TODO убрать токены? для браузера куки и заголовки
      content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokensResponse'
    400-BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    401-Unauthorized:
      description: Invalid or missing access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    404-NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    409-Conflict:
      description: Conflict - object already exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    422-UnprocessableEntity:
      description: Invalid parameter value
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  parameters:
    UserIdParam:
      name: user_id
      in: path
      description: UUID of user
      required: true
      schema:
        type: string
        format: uuid
...
paths:
...
  /auth/v1/refresh:
    post:
      tags:
        - auth
      summary: Get new access and refresh tokens
      description: Get new access and refresh tokens
      operationId: refresh
      parameters:
        - in: cookie
          name: refresh_token_cookie
          schema:
            type: string
        - in: header
          name: X-CSRF-TOKEN
          schema:
            type: string
      # TODO добавить requestBody (для мобильных приложений)
      responses:
        '200':
          $ref: '#/components/responses/200-JWTTokens'
        '401':
          $ref: '#/components/responses/401-Unauthorized'
        '422':
          $ref: '#/components/responses/422-UnprocessableEntity'
...
  /auth/v1/roles:
    get:
      tags:
        - roles
      summary: Get list of all roles
      description: Get list of all roles
      operationId: get_roles
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolesList'
        '401':
          $ref: '#/components/responses/401-Unauthorized'
    post:
      tags:
        - roles
      summary: Create a new role
      description: Create a new role
      operationId: create_role
      security:
        - bearerAuth: [ ]
      requestBody:
        description: Data for creating a new role
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleNew'
      responses:
        '201':
          description: A new role created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleNew'
        '400':
          $ref: '#/components/responses/400-BadRequest'
        '401':
          $ref: '#/components/responses/401-Unauthorized'
        '409':
          $ref: '#/components/responses/409-Conflict'
...
  /auth/v1/users/{user_id}/roles:
    parameters:
      - $ref: '#/components/parameters/UserIdParam'
    get:
      tags:
        - users
      summary: Get list of user roles
      description: Get list of user roles
      operationId: get_user_roles
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolesList'
        '401':
          $ref: '#/components/responses/401-Unauthorized'
        '404':
          $ref: '#/components/responses/404-NotFound'
        '422':
          $ref: '#/components/responses/422-UnprocessableEntity'
...
  /auth/v1/users/{user_id}/history:
    parameters:
      - $ref: '#/components/parameters/UserIdParam'
    get:
      tags:
        - users
      summary: Get user actions history
      description: Get user actions history
      operationId: get_user_history
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of action records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserActionsList'
        '401':
          $ref: '#/components/responses/401-Unauthorized'
        '404':
          $ref: '#/components/responses/404-NotFound'
        '422':
          $ref: '#/components/responses/422-UnprocessableEntity'
  /auth/v1/users/{user_id}/sessions:
    parameters:
      - $ref: '#/components/parameters/UserIdParam'
    get:
      tags:
        - users
      summary: Get active sessions of user
      description: Get active sessions of user
      operationId: get_user_sessions
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSessionsList'
        '401':
          $ref: '#/components/responses/401-Unauthorized'
        '404':
          $ref: '#/components/responses/404-NotFound'
        '422':
          $ref: '#/components/responses/422-UnprocessableEntity'
...